refractory the code structure and add function decomposition
