+++ scripts/gen-single-file.cmake
@@ -13,6 +13,7 @@ loadfile(${my_include_folder}/internal/console.h console)
 loadfile(${my_include_folder}/internal/debugbreak.h debugbreak)
 loadfile(${my_include_folder}/internal/decomposition.h decomposition)
 loadfile(${my_include_folder}/internal/threadsafe.h threadsafe)
+loadfile(${my_include_folder}/internal/rng.h rng)
 
 loadfile(${my_include_folder}/assert.h assert)
 loadfile(${my_include_folder}/benchmark.h benchmark)
@@ -24,6 +25,7 @@ loadfile(${my_include_folder}/log.h log)
 loadfile(${my_include_folder}/table.h table)
 loadfile(${my_include_folder}/unittest.h unittest)
 
+loadfile(${my_src_folder}/rng.cpp rng_cpp)
 loadfile(${my_src_folder}/benchmark.cpp benchmark_cpp)
 loadfile(${my_src_folder}/color.cpp color_cpp)
 loadfile(${my_src_folder}/print.cpp print_cpp)
@@ -33,11 +35,11 @@ loadfile(${my_src_folder}/table.cpp table_cpp)
 loadfile(${my_src_folder}/unittest.cpp unittest_cpp)
 
 
-file(WRITE  zeroerr.hpp "// ====================================================================== lgtm [cpp/missing-header-guard]\n")
+file(WRITE  zeroerr.hpp "// ======================================================================\n")
 file(APPEND zeroerr.hpp "// == DO NOT MODIFY THIS FILE BY HAND - IT IS AUTO GENERATED BY CMAKE! ==\n")
 file(APPEND zeroerr.hpp "// ======================================================================\n")
-file(APPEND zeroerr.hpp "${config}\n${color}\n${console}\n${debugbreak}\n${threadsafe}\n${print}\n${decomposition}\n")
+file(APPEND zeroerr.hpp "${config}\n${color}\n${console}\n${debugbreak}\n${threadsafe}\n${print}\n${decomposition}\n${rng}\n")
 file(APPEND zeroerr.hpp "${benchmark}\n${assert}\n${dbg}\n${format}\n${log}\n${table}\n${profiler}\n${unittest}\n")
 file(APPEND zeroerr.hpp "#ifdef ZEROERR_IMPLEMENTATION\n")
-file(APPEND zeroerr.hpp "${benchmark_cpp}\n${color_cpp}\n${print_cpp}\n${console_cpp}\n${log_cpp}\n${table_cpp}\n${unittest_cpp}\n")
+file(APPEND zeroerr.hpp "${benchmark_cpp}\n${rng_cpp}\n${color_cpp}\n${print_cpp}\n${console_cpp}\n${log_cpp}\n${table_cpp}\n${unittest_cpp}\n")
 file(APPEND zeroerr.hpp "#endif // ZEROERR_IMPLEMENTATION\n")
