+++ test/fuzz_test.cpp
@@ -1,11 +1,13 @@
 #include "zeroerr/fuzztest.h"
+#include "zeroerr/unittest.h"
+#include "zeroerr/assert.h"
 
 #include <string>
 
 using namespace zeroerr;
 
 FUZZ_TEST_CASE("fuzz_test") {
-    FUZZ_FUNC([](int k, std::string num) {
+    FUZZ_FUNC([=](int k, std::string num) {
         int t = atoi(num.c_str());
         REQUIRE(k == t);
     }).WithDomains(
